refactor: RateDto에 단어 나이도 추가
