+++ mobidic_spring/src/main/java/com/kimtaeyang/mobidic/dto/RateDto.java
@@ -1,6 +1,7 @@
 package com.kimtaeyang.mobidic.dto;
 
 import com.kimtaeyang.mobidic.entity.Rate;
+import com.kimtaeyang.mobidic.type.Difficulty;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -17,13 +18,15 @@ public class RateDto {
     private int correctCount;
     private int incorrectCount;
     private int isLearned;
+    private Difficulty difficulty;
 
-    public static RateDto fromEntity(Rate rate) {
+    public static RateDto fromEntity(Rate rate, Difficulty difficulty) {
         return RateDto.builder()
                 .wordId(rate.getWordId())
                 .correctCount(rate.getCorrectCount())
                 .incorrectCount(rate.getIncorrectCount())
                 .isLearned(rate.getIsLearned())
+                .difficulty(difficulty)
                 .build();
     }
 }
