+++ mobidic_spring/src/main/java/com/kimtaeyang/mobidic/service/RateService.java
@@ -8,6 +8,7 @@ import com.kimtaeyang.mobidic.exception.ApiException;
 import com.kimtaeyang.mobidic.repository.RateRepository;
 import com.kimtaeyang.mobidic.repository.VocabRepository;
 import com.kimtaeyang.mobidic.repository.WordRepository;
+import com.kimtaeyang.mobidic.type.Difficulty;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -34,7 +35,7 @@ public class RateService {
         Rate rate = rateRepository.findRateByWord(word)
                 .orElseThrow(() -> new ApiException(NO_RATE));
 
-        return RateDto.fromEntity(rate);
+        return RateDto.fromEntity(rate, getDifficulty(rate.getCorrectCount(), rate.getIncorrectCount()));
     }
 
     @PreAuthorize("@vocabAccessHandler.ownershipCheck(#vocabId)")
@@ -80,5 +81,31 @@ public class RateService {
         Word word = wordRepository.findById(wordId)
                 .orElseThrow(() -> new ApiException(NO_WORD));
         rateRepository.increaseIncorrectCount(word);
+    }
+
+    private Difficulty getDifficulty(Integer correct, Integer incorrect) {
+        double diff = calcDifficultyRatio(correct, incorrect);
+
+        if (diff < 0.3) {
+            return Difficulty.EASY;
+        } else if (diff > 0.7) {
+            return Difficulty.HARD;
+        }
+
+        return Difficulty.NORMAL;
+    }
+
+    private double calcDifficultyRatio(Integer correct, Integer incorrect) {
+        /*
+            난이도 함수 : -0.04correct + 0.05incorrect + 0.5
+        */
+        double diff = (-0.04 * correct) + (0.05 * incorrect) + 0.5;
+        if (diff > 1) {
+            diff = 1;
+        } else if (diff < 0) {
+            diff = 0;
+        }
+
+        return diff;
     }
 }
