+++ mobidic_spring/src/main/java/com/kimtaeyang/mobidic/service/WordService.java
@@ -9,7 +9,6 @@ import com.kimtaeyang.mobidic.exception.ApiException;
 import com.kimtaeyang.mobidic.repository.RateRepository;
 import com.kimtaeyang.mobidic.repository.VocabRepository;
 import com.kimtaeyang.mobidic.repository.WordRepository;
-import com.kimtaeyang.mobidic.type.Difficulty;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -96,31 +95,5 @@ public class WordService {
         wordRepository.delete(word);
 
         return WordDto.fromEntity(word);
-    }
-
-    private Difficulty getDifficulty(Integer correct, Integer incorrect) {
-        double diff = calcDifficultyRatio(correct, incorrect);
-
-        if (diff < 0.3) {
-            return Difficulty.EASY;
-        } else if (diff > 0.7) {
-            return Difficulty.HARD;
-        }
-
-        return Difficulty.NORMAL;
-    }
-
-    private double calcDifficultyRatio(Integer correct, Integer incorrect) {
-        /*
-            난이도 함수 : -0.04correct + 0.05incorrect + 0.5
-        */
-        double diff = (-0.04 * correct) + (0.05 * incorrect) + 0.5;
-        if (diff > 1) {
-            diff = 1;
-        } else if (diff < 0) {
-            diff = 0;
-        }
-
-        return diff;
     }
 }
